Move 'get_files_for_linting' into 'script_utils'.
Addresses:
https://github.com/GoogleCloudPlatform/google-cloud-python/pull/2486#issuecomment-251246970
